pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "research-system-javanese-profile"
version = "0.1.0"
description = "Standalone repo for research_system.commands.javanese_profile."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-docx>=0.8.11",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
javanese-profile = "research_system.commands.javanese_profile:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-p no:cacheprovider"
